Time: 40 ms (72.20%), Space: 90.2 MB (9.79%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cpp
@@ -31,7 +31,7 @@ public:
         if (len % 2 == 1) {
             ret = nums[lenHalf];
         } else {
-            ret = ((double)nums[lenHalf] + nums[lenHalf - 1]) / 2;
+            ret = (nums[lenHalf] + nums[lenHalf - 1]) / 2.0;
         }
         return ret;
     }
